Fix params storage. Ignore _Queue dir.
The params were being stored as a string. When this was being
JSON-serialised later it was causing it to be all escaped. So now
actually store it as key-value pairs, so serialising works correctly.
Move DeleteDirectory functionality to Utils.cpp/h. Ignore the _Queue
directory when looking for changes to the Root directory.